Add global urls to make changes to them easier in the future (if they do change)
Also formatted all files in front end folder
Edit backend endpoitns as well to be more simplistic
