+++ InsomniaEndpoints.json
@@ -1,0 +1,1 @@
+{"_type":"export","__export_format":4,"__export_date":"2024-09-13T23:21:34.490Z","__export_source":"insomnia.desktop.app:v10.0.0","resources":[{"_id":"req_8ea018579c754acf9b9710184a00fbca","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265937524,"created":1725056851701,"url":"http://localhost:3001/auth/login","name":"Log In","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"username\": \"jakob-shavinski\",\n\t\"password\": \"password\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056841284,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"fld_629251bcef5a460194ac7f7a5ba35fd6","parentId":"wrk_9d63c6061c1e400e8911e73eea07ee5d","modified":1726264863433,"created":1725056837486,"name":"auth","description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1725056837486,"_type":"request_group"},{"_id":"wrk_9d63c6061c1e400e8911e73eea07ee5d","parentId":null,"modified":1724434707688,"created":1724434707688,"name":"LiftLog","description":"","scope":"collection","_type":"workspace"},{"_id":"req_225782757cd9461398a9e470ff264e26","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265547552,"created":1724434717275,"url":"http://localhost:3001/auth/signup","name":"Sign Up","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"firstName\": \"jakob\",\n  \"lastName\": \"shavinski\",\n  \"heightFeet\": 5,\n  \"heightInches\": 9,\n  \"weight\": 162,\n  \"bodyType\": \"ectomorph\",\n  \"goal\": \"gain weight\",\n  \"username\": \"jp23shavin\",\n  \"email\": \"test54@gmail.com\",\n  \"password\": \"password\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056841184,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_6efc53653afb408c9856529b21405d60","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265791577,"created":1726265558322,"url":"http://localhost:3001/auth/part-1","name":"Sign up part 1","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"firstName\": \"jakob\",\n  \"lastName\": \"shavinski\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056841134,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_f49f0c9fdc92479683b2ceabe4b10007","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265642260,"created":1726265586060,"url":"http://localhost:3001/auth/part-2","name":"Sign up part 2","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n  \"heightFeet\": 5,\n  \"heightInches\": 9,\n  \"weight\": 162\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056841084,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_8259ccc657774fd5b0c6a00b32fb5c93","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265686868,"created":1726265589609,"url":"http://localhost:3001/auth/part-3","name":"Sign up part 3","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n  \"body\": \"ectomorph\"\n }"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056841034,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_daee91d42731491394c52fd3baac8d4a","parentId":"fld_629251bcef5a460194ac7f7a5ba35fd6","modified":1726265715851,"created":1726265597251,"url":"http://localhost:3001/auth/part-4","name":"Sign up part 4","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n  \"goal\": \"gain weight\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json"},{"name":"User-Agent","value":"insomnia/9.3.3"}],"authentication":{},"metaSortKey":-1725056840984,"isPrivate":false,"pathParameters":[],"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"env_15a48cf84bc33fff450f6a94af42c5ad9da38a53","parentId":"wrk_9d63c6061c1e400e8911e73eea07ee5d","modified":1724434707690,"created":1724434707690,"name":"Base Environment","data":{},"dataPropertyOrder":null,"color":null,"isPrivate":false,"metaSortKey":1724434707690,"_type":"environment"},{"_id":"jar_15a48cf84bc33fff450f6a94af42c5ad9da38a53","parentId":"wrk_9d63c6061c1e400e8911e73eea07ee5d","modified":1724434707692,"created":1724434707692,"name":"Default Jar","cookies":[],"_type":"cookie_jar"}]}
