+++ backend-lift-log/src/app.ts
@@ -5,6 +5,8 @@ import { PORT } from './config';
 
 // ROUTE IMPORTS 
 import userRouter from './routes/userRoutes';
+import authRouter from './routes/authRoutes';
+
 
 const app = express();
 
@@ -13,6 +15,7 @@ app.use(cors());
 
 // ROUTES
 app.use('/users', userRouter);
+app.use('/auth', authRouter);
 
 app.get('/', (req, res) => {
     res.send('Hello World');
