+++ backend-lift-log/src/controllers/authController.ts
@@ -2,25 +2,6 @@ import { Request, Response } from 'express';
 import User from '../models/user';
 import { createToken } from '../helpers/tokens';
 
-
-export const getAllUsersTest = async(req: Request, res: Response) => {
-    try {
-        const result = await User.getAllUsers();
-        res.json(result);
-    } catch (err) {
-        res.status(400).send('Cannot find this source');
-    }
-}
-
-export const getSingleUserData = async(req: Request, res: Response) => {
-    try {
-        const user = await User.getSingleUserData(req.params.username);
-        return res.json({user});
-    } catch (err) {
-        res.status(400).send("User not found");
-    }
-}
-
 export const registerUserPart1 = async (req: Request, res: Response) => {
     try {
         const info = await User.validatePartOneForm(req.body);
@@ -32,8 +13,8 @@ export const registerUserPart1 = async (req: Request, res: Response) => {
 
 export const registerUserPart2 = async (req: Request, res: Response) => {
     try {
-        await User.validatePartTwoForm(req.body);
-        res.status(200).json({ message: "Part 2 success" })
+        const info = await User.validatePartTwoForm(req.body);
+        res.status(200).json({ message: "Part 2 success", info })
     } catch (error: any) {
         res.status(400).json({ error: 'Invalid data', message: error.messages });
     }
@@ -41,8 +22,8 @@ export const registerUserPart2 = async (req: Request, res: Response) => {
 
 export const registerUserPart3 = async (req: Request, res: Response) => {
     try {
-        await User.validatePartThreeForm(req.body);
-        res.status(200).json({ message: "Part 3 success" })
+        const info = await User.validatePartThreeForm(req.body);
+        res.status(200).json({ message: "Part 3 success", info})
     } catch (error: any) {
         res.status(400).json({ error: 'Invalid data', message: error.messages });
     }
@@ -50,8 +31,8 @@ export const registerUserPart3 = async (req: Request, res: Response) => {
 
 export const registerUserPart4 = async (req: Request, res: Response) => {
     try {
-        await User.validatePartFourForm(req.body);
-        res.status(200).json({ message: "Part 4 success" })
+        const info = await User.validatePartFourForm(req.body);
+        res.status(200).json({ message: "Part 4 success", info })
     } catch (error: any) {
         res.status(400).json({ error: 'Invalid data', messages: error.messages });
     }
@@ -59,9 +40,9 @@ export const registerUserPart4 = async (req: Request, res: Response) => {
 
 export const signup = async (req: Request, res: Response) => {
     try {
-        const user = await User.signup({...req.body, isAdmin: false});
+        const user = await User.signup({ ...req.body, isAdmin: false });
         const token = createToken(user);
-        return res.status(200).json({ message: "Sign up success", token})
+        return res.status(200).json({ message: "Sign up success", token })
     } catch (error: any) {
         res.status(400).json({ error: 'Invalid data', messages: error.messages });
     }
@@ -69,10 +50,10 @@ export const signup = async (req: Request, res: Response) => {
 
 export const login = async (req: Request, res: Response) => {
     try {
-        const user = await User.login({...req.body});
+        const user = await User.login({ ...req.body });
         console.log(req.body)
         const token = createToken(user);
-        return res.status(200).json({ message: "Log in success", token})
+        return res.status(200).json({ message: "Log in success", token })
     } catch (error: any) {
         res.status(400).json({ error: 'Invalid data', messages: error.messages });
     }
