+++ frontend-lift-log/src/api/liftLogApi.ts
@@ -1,4 +1,5 @@
 import axios from "axios";
+import authEndpoints from "../constants/urls";
 
 const BASE_URL = `http://localhost:3001`
 
@@ -15,7 +16,6 @@ export type signUpData = {
     password: string;
     isAdmin?: boolean;
 };
-
 export type loginData = {
     username: string,
     password: string
@@ -36,32 +36,33 @@ export class LiftLogApi {
     }
 
     // create a sign up method
-        static async signup(formData: signUpData) {
-            try {
-                const res = await axios.post(`${BASE_URL}/users/create/account/signup`, formData);
-                return res.data.token;
-            } catch (error) {
-                // TODO: Make this into a helper function 
-                const isAxiosError = axios.isAxiosError(error);
-                const messages = isAxiosError ? error?.response?.data.messages : "An unexpected error occured";
-                throw messages;
-            }
+    static async signup(formData: signUpData) {
+        try {
+            const res = await axios.post(`${BASE_URL}/auth/${authEndpoints.signupPath}`, formData);
+            return res.data.token;
+        } catch (error) {
+            // TODO: Make this into a helper function 
+            const isAxiosError = axios.isAxiosError(error);
+            const messages = isAxiosError ? error?.response?.data.messages : "An unexpected error occured";
+            throw messages;
         }
+    }
 
     // create a login method 
-        static async login(formData: loginData) {
-            try {
-                const res = await axios.post(`${BASE_URL}/users/account/login`, formData);
-                return res.data.token;
-            } catch (error) {
-                const isAxiosError = axios.isAxiosError(error);
-                const messages = isAxiosError ? error?.response?.data.messages : "An unexpected error occured";
-                throw messages;
-            }
+    static async login(formData: loginData) {
+        try {
+            const res = await axios.post(`${BASE_URL}/auth/${authEndpoints.loginPath}`, formData);
+            return res.data.token;
+        } catch (error) {
+            const isAxiosError = axios.isAxiosError(error);
+            const messages = isAxiosError ? error?.response?.data.messages : "An unexpected error occured";
+            throw messages;
         }
+    }
 
-        static logout() {
-            this.token = "";
-        }
+    // Log out method
+    static logout() {
+        this.token = "";
+    }
 
 }
