+++ frontend-lift-log/src/components/forms/MultiStepSignup/Part4.tsx
@@ -1,38 +1,35 @@
 import React, { FC, useState } from "react";
-
-import ectomorphBody from "/images/bodyForm/ectomorphBody.jpg"
-import mesomorphBody from "/images/bodyForm/mesomorphBody.jpg"
-import endomorphBody from "/images/bodyForm/endomorphBody.jpg"
-
-import BodyTypeContainer from "./Step3Container";
 import { useNavigate } from "react-router-dom";
+
+import GoalContainer from "./Part4Container";
+import authEndpoints from "../../../constants/urls";
 // import { useFormProgessBar } from "./SignupForm";
 
 interface FormData {
-    body: string,
+    goal: string,
 }
-
 interface ErrorData {
-    body?: string
+    goal?: string,
 }
 
-// DO NOT CHANGE, includes body type, img path, and description
-const BODY_TYPE_DETAILS: string[][] = [
-    ["Ectomorph", ectomorphBody, "Typically a person with a slimmer and leaner body. Naturally less body fat and muscle overall."],
-    ["Mesomorph", mesomorphBody, "Typically a person with a naturally more athletic and strong body."],
-    ["Endomorph", endomorphBody, "Typically a person that contains more stored fat, more muscle, and gains weight easily."]
-]
+const GOAL_OPTIONS = [
+    "Lose weight",
+    "Maintain current weight",
+    "Gain weight",
+    "Gain muscle",
+    "Become more active"
+];
 
-const Step3: FC = () => {
+const Part4: FC = () => {
     const navigate = useNavigate();
     // const { nextStep, prevStep } = useFormProgessBar();
 
     const [formData, setFormData] = useState<FormData>({
-        body: sessionStorage.getItem('body') ?? "",
-    });
+        goal: sessionStorage.getItem('goal') ?? "",
+    })
 
     const [errors, setErrors] = useState<ErrorData>({
-        body: "",
+        goal: "",
     })
 
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -42,38 +39,37 @@ const Step3: FC = () => {
     }
 
     const handleBack = () => {
-        // prevStep();
-        navigate("/users/create/account/part-2-height-weight");
+        navigate(`/auth/${authEndpoints.part3Path}`);
     }
 
     const validateForm = (e: React.FormEvent) => {
         e.preventDefault();
 
         const newErrors: ErrorData = {}
-        if (!formData.body) newErrors.body = "Please select one body type"
+        if (!formData.goal) newErrors.goal = "Please choose one goal"
 
         setErrors(newErrors);
         if (Object.keys(newErrors).length > 0) return;
 
-        sessionStorage.setItem("body", formData.body);
+        sessionStorage.setItem("goal", formData.goal);
 
-        navigate("/users/create/account/part-4-goal");
+        navigate(`/auth/${authEndpoints.signupPath}`);
     }
 
     return (
-
         <form
             onSubmit={validateForm}
             className="flex flex-col w-full p-4 md:shadow-custom"
-            data-testid="body-type-form">
+            data-testid="goal-form">
 
-            <h1 className="font-bold text-left mt-3 mb-3 text-lg">What type of body type do you have?</h1>
+            <h1 className="font-bold text-left mt-3 mb-3 text-lg">Personal Goals</h1>
+            <small className="mb-1">Please select one.</small>
 
-            {BODY_TYPE_DETAILS.map((type, i) => {
-                return <BodyTypeContainer currentValue={formData.body} body={type[0]} imagePath={type[1]} info={type[2]} handleChange={handleChange} key={i} />;
+            {GOAL_OPTIONS.map((goal, i) => {
+                return <GoalContainer currentValue={formData.goal} goal={goal} handleChange={handleChange} key={i} />;
             })}
 
-            {errors.body && <span className="text-red-500 ml-2 text-sm">{errors.body}</span>}
+            {errors.goal && <span className="text-red-500 ml-2 text-sm">{errors.goal}</span>}
 
             <div className="flex gap-5 mt-8">
                 <button
@@ -86,9 +82,10 @@ const Step3: FC = () => {
                     type="submit"
                     className="w-1/2 p-3 bg-[#00df9a] rounded-md hover:bg-[#10B981] text-white font-bold text-xl"
                 >Next</button>
+
             </div>
         </form>
     )
 }
 
-export default Step3;
+export default Part4;
