+++ frontend-lift-log/src/constants/urls.ts
@@ -1,0 +1,12 @@
+const BASE_URL = `http://localhost:3001`;
+
+const authEndpoints = {
+    loginPath: `login`,
+    part1Path: `part-1`,
+    part2Path: `part-2`,
+    part3Path: `part-3`,
+    part4Path: `part-4`,
+    signupPath: `signup`
+}
+
+export default authEndpoints;
